requirements.txt
us
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
boto3==1.18.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0
django-allauth
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
drf-yasg==1.21.7
filelock==3.13.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.1.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
# urllib3==2.1.0
uWSGI==2.0.20
virtualenv==20.25.0
whitenoise==6.6.0
yarl==1.9.4
django-prometheus